creando capa orden y orden_detalles

+++ server/src/main/java/c1521mjavaangular/ecotienda/repositories/ProductoRepository.java
@@ -11,5 +11,5 @@ import java.util.Optional;
 public interface ProductoRepository extends JpaRepository<Productos, Integer>{
 
     @EntityGraph(attributePaths = "categorias")
-    Optional<Productos> findById(Integer id);
+    Optional<Productos> findById(Long id);
 }
